Use floatField for prices, for pyYAML to convert to type "number"

+++ order/models.py
@@ -6,13 +6,13 @@ from django.db import models
 class Order(models.Model):
     detail = models.TextField()
     clientName = models.TextField()
-    price = models.DecimalField(max_digits=9, decimal_places=2)
+    price = models.FloatField()
     isPaid = models.BooleanField(default=False)
 
 
 class Payment(models.Model):
     order = models.OneToOneField(Order, on_delete=models.CASCADE)
-    price = models.DecimalField(max_digits=9, decimal_places=2)
+    price = models.FloatField()
     isSuccessful = models.BooleanField(default=False)
 
 
@@ -20,4 +20,4 @@ class MenuItem(models.Model):
     name = models.TextField()
     MENU_ITEM_TYPES = [('coffee', 'Coffee'), ('addition', 'Addition')]
     type = models.CharField(max_length=8, choices=MENU_ITEM_TYPES)
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.FloatField()
